+++ packages/bluesky/lib/src/entities/converter/reply_post_converter.dart
@@ -35,7 +35,7 @@ class ReplyPostConverter
       return ReplyPost.record(
         data: Post.fromJson(json),
       );
-    } on Exception {
+    } on UnsupportedError {
       return ReplyPost.unknown(data: json);
     }
   }