+++ packages/bluesky_cli/lib/src/command/feed/feed_generators.dart
@@ -1,0 +1,43 @@
+// Copyright 2023 Shinya Kato. All rights reserved.
+// Redistribution and use in source and binary forms, with or without
+// modification, are permitted provided the conditions.
+
+import 'package:at_uri/at_uri.dart';
+import 'package:xrpc/xrpc.dart' as xrpc;
+
+import '../query_command.dart';
+
+/// The command for `app.bsky.feed.getFeedGenerators`.
+class FeedGeneratorsCommand extends QueryCommand {
+  /// Returns the new instance of [FeedGeneratorsCommand].
+  FeedGeneratorsCommand() {
+    argParser.addOption(
+      'uris',
+      help: "AT URI of generator to be retrieved in CSV format.",
+      defaultsTo: '',
+    );
+  }
+
+  @override
+  final String name = 'feed-generators';
+
+  @override
+  final String description = 'Show the specific feed generators.';
+
+  @override
+  final String invocation = 'bsky feed-generators [uris]';
+
+  @override
+  xrpc.NSID get methodId => xrpc.NSID.create(
+        'feed.bsky.app',
+        'getFeedGenerators',
+      );
+
+  @override
+  Map<String, dynamic>? get parameters => {
+        'feeds': argResults!['uris']
+            .split(',')
+            .map((e) => AtUri.parse(e).toString())
+            .toList(),
+      };
+}